对接公众号接口，授权和获取用户信息等

+++ src/main/java/com/lulu/reservation/repository/SmsRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @version 2.0.1 创建时间: 2018/7/7 下午4:51
  * <p>
  * 类说明：
- *     短信Dao
+ *     短信Repository
  */
-public interface SmsRepository extends JpaRepository<Sms, Integer> {
+public interface SmsRepository extends JpaRepository<Sms, Long> {
 }
